+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/annotation/RequestMapping.java
@@ -1,0 +1,18 @@
+package cn.haitaoss.tinyspringmvc.framework.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-24 18:45
+ *
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface RequestMapping {
+    String value() default "";
+}