+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/HandleInterceptor.java
@@ -11,9 +11,30 @@ import javax.servlet.http.HttpServletResponse;
  * date 2021-04-24 17:05
  */
 public interface HandleInterceptor extends MethodInterceptor {
-    boolean preHandle(HttpServletRequest var1, HttpServletResponse var2, Object var3) throws Exception;
+    /**
+     * 该方法在请求处理之前调用，返回true表示交给下一个拦截器，返回false表示到此为止
+     * @author haitao.chen
+     * email
+     * date 2021/4/24 7:10 下午
+     * @param request
+     * @param response
+     * @param handler
+     * @return boolean
+     */
+    boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception;
 
-    // void postHandle(HttpServletRequest var1, HttpServletResponse var2, Object var3, ModelAndView var4) throws Exception;
+    /**
+     * 视图返回之后，渲染之前被调用
+     * @author haitao.chen
+     * email
+     * date 2021/4/24 7:10 下午
+     * @param request
+     * @param response
+     * @param handler
+     * @param modelAndView
+     */
+    void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception;
 
-    // void afterCompletion(HttpServletRequest var1, HttpServletResponse var2, Object var3, Exception var4) throws Exception;
+    void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception;
+
 }