+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/ModelAndView.java
@@ -1,0 +1,10 @@
+package cn.haitaoss.tinyspringmvc.framework.handlerMapping;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-24 19:10
+ *
+ */
+public class ModelAndView {
+}