+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/servlet/DispatcherServlet.java
@@ -2,11 +2,16 @@ package cn.haitaoss.tinyspringmvc.framework.servlet;
 
 import cn.haitaoss.tinyioc.context.ApplicationContext;
 import cn.haitaoss.tinyioc.context.ClassPathXmlApplicationContext;
-import cn.haitaoss.tinyspringmvc.webappProject.controller.PersonController;
+import cn.haitaoss.tinyspringmvc.framework.handlerMapping.AnnotationHandleMapping;
+import cn.haitaoss.tinyspringmvc.framework.handlerMapping.HandlerExecutionChain;
+import cn.haitaoss.tinyspringmvc.framework.handlerMapping.HandlerMapping;
+import cn.haitaoss.tinyspringmvc.framework.handlerMapping.RequestMappingHandler;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author haitao.chen
@@ -15,12 +20,17 @@ import javax.servlet.http.HttpServlet;
  *
  */
 public class DispatcherServlet extends HttpServlet {
-    public DispatcherServlet() {
-        System.out.println("DispatcherServlet...constructor");
-    }
+    private ApplicationContext mvcContext;
+    private HandlerMapping handlerMapping;
 
     @Override
     public void init() throws ServletException {
+        doInit();
+        // 默认使用AnnotationHandlerMapping来处理请求
+        handlerMapping = new AnnotationHandleMapping(mvcContext);
+    }
+
+    private void doInit() throws ServletException {
         super.init();
         System.out.println("DispatcherServlet...init\tspring子容器(mvc)");
         // 获取mvc配置文件
@@ -33,19 +43,32 @@ public class DispatcherServlet extends HttpServlet {
         }
         // 获取全局上下文
         ServletContext servletContext = this.getServletContext();
-        ApplicationContext mvcContext = null;
         try {
             // 获取父容器
             ApplicationContext springContext = (ApplicationContext) servletContext.getAttribute("springContext");
             // 创建mvc的容器，里面整合和spring作为父容器
             mvcContext = new ClassPathXmlApplicationContext(springContext, mvcXmlPath);
-
-            // 测试
-            PersonController personController = (PersonController) mvcContext.getBean("personController");
-            System.out.println(personController.getPersonService());
-
         } catch (Exception e) {
-            // e.printStackTrace();
+            e.printStackTrace();
         }
+    }
+
+    @Override
+    protected void service(HttpServletRequest req, HttpServletResponse resp) {
+        try {
+            doDispatch(req, resp);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void doDispatch(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        // 找到请求对应的handler
+        HandlerExecutionChain handlerExecutionChain = handlerMapping.getHandler(req);
+
+        // 执行目标方法
+        RequestMappingHandler handler = handlerExecutionChain.getHandler();
+        // 至于如何传参，就是HandlerAdapter的事情了
+        handler.getMethod().invoke(handler.getBean(), null);
     }
 }