+++ src/main/java/cn/haitaoss/tinyspringmvc/webappProject/controller/PersonController.java
@@ -2,6 +2,7 @@ package cn.haitaoss.tinyspringmvc.webappProject.controller;
 
 import cn.haitaoss.tinyioc.beans.annotation.Autowired;
 import cn.haitaoss.tinyioc.beans.annotation.Controller;
+import cn.haitaoss.tinyspringmvc.framework.annotation.RequestMapping;
 import cn.haitaoss.tinyspringmvc.webappProject.service.PersonService;
 
 /**
@@ -11,9 +12,27 @@ import cn.haitaoss.tinyspringmvc.webappProject.service.PersonService;
  *
  */
 @Controller
+@RequestMapping("/mvc/person/")
 public class PersonController {
     @Autowired
     private PersonService personService;
+
+    @RequestMapping("eat.do")
+    public void eating() {
+        System.out.println("I am eating!");
+    }
+
+    /**
+     * 这个还不能测试，没有实现参数的分析
+     * @author haitao.chen
+     * email
+     * date 2021/4/24 8:26 下午
+     * @param age
+     */
+    @RequestMapping("speak.do")
+    public void speak(int age) {
+        System.out.println("i am " + age + "years old");
+    }
 
     public PersonService getPersonService() {
         return personService;